预付费出院结清，收费

+++ ruoyi-system/src/main/java/com/ruoyi/system/fantang/domain/FtPrepaymentDao.java
@@ -74,7 +74,7 @@ public class FtPrepaymentDao implements Serializable {
      * 结算标志
      */
     @Excel(name = "结算标志")
-    private Long settlementFlag;
+    private Integer settlementFlag;
 
     /**
      * 预付费金额
